# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Registre des erreurs API
  version: 1.0.0
  description: Contrat API officiel de la V1.
  license:
    name: Propriétaire — usage interne uniquement
    identifier: LicenseRef-Proprietary
servers:
- url: /api
security:
- CloudflareAccessJWT: []
...
paths:
...
  /issues:
    get:
      tags:
      - Issues
      operationId: listIssues
      summary: Lister les dossiers avec filtres et pagination.
      parameters:
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: q
        in: query
        schema:
          type: string
          minLength: 2
          maxLength: 40
        x-maxUtf8Bytes: 160
      - name: status
        in: query
        schema:
          type: array
          items:
            $ref: '#/components/schemas/IssueStatus'
        style: form
        explode: true
      - name: priority
        in: query
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Priority'
        style: form
        explode: true
      - $ref: '#/components/parameters/LocationIdQuery'
      - $ref: '#/components/parameters/DepartmentIdQuery'
      - $ref: '#/components/parameters/CategoryIdQuery'
      - name: ownerUserId
        in: query
        schema:
          type: integer
          minimum: 1
      - name: errorActorUserId
        in: query
        description: Employé concerné par l'erreur.
        schema:
          type: integer
          minimum: 1
      - name: sort
        in: query
        schema:
          type: string
          enum:
          - newest
          - oldest
          - priority
          - dueDate
          default: newest
      - name: from
        in: query
        schema:
          type: string
          format: date
      - name: to
        in: query
        schema:
          type: string
          format: date
      - name: overdue
        in: query
        schema:
          type: boolean
      - name: effectivenessStatus
        in: query
        schema:
          $ref: '#/components/schemas/EffectivenessStatus'
      - $ref: '#/components/parameters/EffectivenessReviewDueBefore'
      responses:
        '200':
          description: Liste paginée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
    post:
      tags:
      - Issues
      operationId: createIssue
      summary: Créer un dossier.
      x-rate-limit: write
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Dossier créé.
          headers:
            ETag:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimited'
  /issues/{publicId}:
    parameters:
    - $ref: '#/components/parameters/PublicId'
    get:
      tags:
      - Issues
      operationId: getIssue
      summary: Récupérer le détail d'un dossier.
      responses:
        '200':
          description: Détail cœur, impacts et actions correctives.
          headers:
            ETag:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
      - Issues
      operationId: updateIssue
      summary: Modifier un dossier.
      x-rate-limit: write
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateIssueRequest'
      responses:
        '200':
          description: Dossier modifié.
          headers:
            ETag:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueDetailResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '428':
          $ref: '#/components/responses/PreconditionRequired'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /analytics/summary:
    get:
      tags:
      - Analytics
      operationId: getAnalyticsSummary
      summary: Récupérer les KPI de synthèse.
      parameters:
      - $ref: '#/components/parameters/DateFrom'
      - $ref: '#/components/parameters/DateTo'
      - $ref: '#/components/parameters/LocationIdQuery'
      - $ref: '#/components/parameters/DepartmentIdQuery'
      - $ref: '#/components/parameters/CategoryIdQuery'
      responses:
        '200':
          description: KPI de synthèse.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /analytics/recurring:
    get:
      tags:
      - Analytics
      operationId: getRecurringIssues
      summary: Récupérer les groupes de récurrences.
      parameters:
      - $ref: '#/components/parameters/LocationIdQuery'
      - $ref: '#/components/parameters/DepartmentIdQuery'
      - $ref: '#/components/parameters/CategoryIdQuery'
      responses:
        '200':
          description: Groupes récurrents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecurringResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      description: Retourne les récurrences locales (location+subcategory) et organisationnelles (subcategory) selon
        seuil 3/90 jours.
  /analytics/effectiveness:
    get:
      tags:
      - Analytics
      operationId: getEffectiveness
      summary: Récupérer l'efficacité des corrections.
      parameters:
      - $ref: '#/components/parameters/DateFrom'
      - $ref: '#/components/parameters/DateTo'
      responses:
        '200':
          description: Efficacité des corrections.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectivenessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
      description: Agrège uniquement issues.effectiveness_status; les actions correctives sont exclues du KPI principal.
  /analytics/errors-by-employee:
    get:
      tags:
      - Analytics
      operationId: getErrorsByEmployee
      summary: Agréger les erreurs par employé et sous-catégorie.
      parameters:
      - $ref: '#/components/parameters/DateFrom'
      - $ref: '#/components/parameters/DateTo'
      - $ref: '#/components/parameters/LocationIdQuery'
      - $ref: '#/components/parameters/DepartmentIdQuery'
      - $ref: '#/components/parameters/CategoryIdQuery'
      responses:
        '200':
          description: Agrégations par employé concerné et sous-catégorie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeErrorStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      description: Réservé aux gestionnaires et administrateurs. Aucun courriel ni texte libre n'est retourné.
...
components:
  securitySchemes:
    CloudflareAccessJWT:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
  parameters:
    PublicId:
      name: publicId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/PublicId'
...
    Cursor:
      name: cursor
      in: query
      schema:
        type:
        - string
        - 'null'
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
    DateFrom:
      name: from
      in: query
      schema:
        type: string
        format: date
    DateTo:
      name: to
      in: query
      schema:
        type: string
        format: date
    LocationIdQuery:
      name: locationId
      in: query
      schema:
        type: integer
        minimum: 1
    DepartmentIdQuery:
      name: departmentId
      in: query
      schema:
        type: integer
        minimum: 1
    CategoryIdQuery:
      name: categoryId
      in: query
      schema:
        type: integer
        minimum: 1
    EffectivenessReviewDueBefore:
      name: effectivenessReviewDueBefore
      in: query
      schema:
        type: string
        format: date
      description: Retourne les dossiers pending dont la date de revue est <= à cette date.
  schemas:
    PublicId:
      type: string
      pattern: ^INC-[0-9]{6,}$
      example: INC-000042
      description: 'Numéro de dossier unique dérivé de l''ID D1 global: INC-{id sur au moins 6 chiffres}. Ce numéro
        n''est ni un compteur annuel ni une date.'
...
    IssueStatus:
      type: string
      enum:
      - new
      - inProgress
      - waiting
      - resolved
    Priority:
      type: string
      enum:
      - normal
      - important
      - urgent
    CauseStatus:
      type: string
      enum:
      - toVerify
      - known
    EffectivenessStatus:
      type: string
      enum:
      - pending
      - effective
      - ineffective
    CorrectiveActionStatus:
      type: string
      enum:
      - todo
      - inProgress
      - waiting
      - done
    PermanentCorrectionType:
      type: string
      enum:
      - procedureUpdate
      - newProcedure
      - training
      - systemConfiguration
      - responsibilityChange
      - additionalCheck
      - supplierProcess
      - noChangeRequired
      - other
    WaitingOn:
      oneOf:
      - type: object
        additionalProperties: false
        required:
        - type
        - userId
        properties:
          type:
            type: string
            const: user
          userId:
            type: integer
            minimum: 1
          label:
            type:
            - string
            - 'null'
            maxLength: 120
      - type: object
        additionalProperties: false
        required:
        - type
        - label
        properties:
          type:
            type: string
            enum:
            - customer
            - supplier
            - other
          userId:
            type:
            - integer
            - 'null'
            minimum: 1
          label:
            type: string
            minLength: 1
            maxLength: 120
      description: 'Une attente doit toujours identifier son objet: utilisateur interne ou libellé externe.'
...
    ImpactInput:
      type: object
      additionalProperties: false
      required:
      - impactTypeId
      properties:
        impactTypeId:
          type: integer
          minimum: 1
        details:
          type:
          - string
          - 'null'
          maxLength: 1000
    Impact:
      allOf:
      - $ref: '#/components/schemas/ImpactInput'
      - type: object
        required:
        - id
        properties:
          id:
            type: integer
            minimum: 1
    CreateIssueRequest:
      type: object
      additionalProperties: false
      required:
      - occurredOn
      - categoryId
      - description
      - priority
      - impacts
      - locationId
      properties:
        occurredOn:
          type: string
          format: date
        locationId:
          type: integer
          minimum: 1
        departmentId:
          type:
          - integer
          - 'null'
          minimum: 1
        categoryId:
          type: integer
          minimum: 1
        subcategoryId:
          type:
          - integer
          - 'null'
          minimum: 1
        description:
          type: string
          minLength: 10
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/Priority'
        impacts:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/ImpactInput'
    UpdateIssueRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        occurredOn:
          type: string
          format: date
        locationId:
          type: integer
          minimum: 1
        departmentId:
          type:
          - integer
          - 'null'
          minimum: 1
        categoryId:
          type: integer
          minimum: 1
        subcategoryId:
          type:
          - integer
          - 'null'
          minimum: 1
        description:
          type: string
          minLength: 10
          maxLength: 5000
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/IssueStatus'
        ownerUserId:
          type:
          - integer
          - 'null'
          minimum: 1
        errorActorUserId:
          type:
          - integer
          - 'null'
          minimum: 1
        dueDate:
          type:
          - string
          - 'null'
          format: date
        causeStatus:
          oneOf:
          - $ref: '#/components/schemas/CauseStatus'
          - type: 'null'
        causeSummary:
          type:
          - string
          - 'null'
          maxLength: 5000
        immediateSolution:
          type:
          - string
          - 'null'
          maxLength: 5000
        permanentCorrectionType:
          oneOf:
          - $ref: '#/components/schemas/PermanentCorrectionType'
          - type: 'null'
        permanentCorrectionSummary:
          type:
          - string
          - 'null'
          maxLength: 5000
        waitingOn:
          oneOf:
          - $ref: '#/components/schemas/WaitingOn'
          - type: 'null'
        finalResult:
          type:
          - string
          - 'null'
          maxLength: 5000
        preventionLearning:
          type:
          - string
          - 'null'
          maxLength: 5000
        effectivenessStatus:
          oneOf:
          - $ref: '#/components/schemas/EffectivenessStatus'
          - type: 'null'
        effectivenessReviewDate:
          type:
          - string
          - 'null'
          format: date
        reopenReason:
          type:
          - string
          - 'null'
          minLength: 5
          maxLength: 1000
        impacts:
          type: array
          minItems: 1
          maxItems: 10
          items:
            $ref: '#/components/schemas/ImpactInput'
    Issue:
      type: object
      required:
      - publicId
      - occurredOn
      - categoryId
      - description
      - priority
      - status
      - rowVersion
      - createdByUserId
      - createdAt
      - updatedAt
      - errorActorUserId
      properties:
        publicId:
          $ref: '#/components/schemas/PublicId'
        occurredOn:
          type: string
          format: date
        locationId:
          type: integer
          minimum: 1
        departmentId:
          type:
          - integer
          - 'null'
        categoryId:
          type: integer
        subcategoryId:
          type:
          - integer
          - 'null'
        description:
          type: string
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/IssueStatus'
        ownerUserId:
          type:
          - integer
          - 'null'
        errorActorUserId:
          type:
          - integer
          - 'null'
        dueDate:
          type:
          - string
          - 'null'
          format: date
        causeStatus:
          oneOf:
          - $ref: '#/components/schemas/CauseStatus'
          - type: 'null'
        causeSummary:
          type:
          - string
          - 'null'
        immediateSolution:
          type:
          - string
          - 'null'
        permanentCorrectionType:
          oneOf:
          - $ref: '#/components/schemas/PermanentCorrectionType'
          - type: 'null'
        permanentCorrectionSummary:
          type:
          - string
          - 'null'
        waitingOn:
          oneOf:
          - $ref: '#/components/schemas/WaitingOn'
          - type: 'null'
        finalResult:
          type:
          - string
          - 'null'
        preventionLearning:
          type:
          - string
          - 'null'
        effectivenessStatus:
          oneOf:
          - $ref: '#/components/schemas/EffectivenessStatus'
          - type: 'null'
        effectivenessReviewDate:
          type:
          - string
          - 'null'
          format: date
        rowVersion:
          type: integer
          minimum: 1
        createdByUserId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        resolvedAt:
          type:
          - string
          - 'null'
          format: date-time
        resolvedByUserId:
          type:
          - integer
          - 'null'
        redactedAt:
          type:
          - string
          - 'null'
          format: date-time
        redactedByUserId:
          type:
          - integer
          - 'null'
        redactionReason:
          type:
          - string
          - 'null'
    CorrectiveAction:
      type: object
      required:
      - id
      - issuePublicId
      - title
      - ownerUserId
      - dueDate
      - status
      - blocksIssueClosure
      - createdAt
      - updatedAt
      properties:
        id:
          type: integer
        issuePublicId:
          $ref: '#/components/schemas/PublicId'
        title:
          type: string
        description:
          type:
          - string
          - 'null'
        ownerUserId:
          type: integer
        dueDate:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/CorrectiveActionStatus'
        blocksIssueClosure:
          type: boolean
        result:
          type:
          - string
          - 'null'
        effectivenessStatus:
          oneOf:
          - $ref: '#/components/schemas/EffectivenessStatus'
          - type: 'null'
        completedAt:
          type:
          - string
          - 'null'
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AnalyticsSummary:
      type: object
      required:
      - open
      - urgent
      - overdue
      - waiting
      - resolved
      - pendingEffectiveness
      properties:
        open:
          type: integer
        urgent:
          type: integer
        overdue:
          type: integer
        waiting:
          type: integer
        resolved:
          type: integer
        pendingEffectiveness:
          type: integer
        averageResolutionHours:
          type:
          - number
          - 'null'
          description: Moyenne en heures calendaires de resolvedAt - createdAt pour les dossiers résolus dans le
            filtre.
    RecurringGroup:
      type: object
      required:
      - scope
      - subcategoryId
      - count
      - windowDays
      properties:
        subcategoryId:
          type: integer
        count:
          type: integer
        windowDays:
          type: integer
        latestIssuePublicId:
          $ref: '#/components/schemas/PublicId'
        scope:
          type: string
          enum:
          - location
          - organization
        locationId:
          type:
          - integer
          - 'null'
          description: Présent pour scope=location; null pour organization.
    Effectiveness:
      type: object
      required:
      - pending
      - effective
      - ineffective
      properties:
        pending:
          type: integer
        effective:
          type: integer
        ineffective:
          type: integer
        effectivenessRate:
          type:
          - number
          - 'null'
          description: effective / (effective + ineffective)
      description: Agrégation basée uniquement sur issues.effectiveness_status. corrective_actions.effectiveness_status
        n'alimente pas ce KPI V1.
    EmployeeErrorStat:
      type: object
      additionalProperties: false
      required:
      - userId
      - displayName
      - active
      - subcategoryId
      - count
      - latestIssuePublicId
      properties:
        userId:
          type: integer
          minimum: 1
        displayName:
          type: string
        active:
          type: boolean
        subcategoryId:
          type: integer
          minimum: 1
        count:
          type: integer
          minimum: 1
        latestIssuePublicId:
          $ref: '#/components/schemas/PublicId'
    ErrorBody:
      type: object
      required:
      - ok
      - error
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required:
          - code
          - message
          - requestId
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
            requestId:
              type: string
...
    IssueResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Issue'
    IssueDetail:
      type: object
      required:
      - issue
      - impacts
      - correctiveActions
      properties:
        issue:
          $ref: '#/components/schemas/Issue'
        impacts:
          type: array
          items:
            $ref: '#/components/schemas/Impact'
        correctiveActions:
          type: array
          items:
            $ref: '#/components/schemas/CorrectiveAction'
    IssueDetailResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/IssueDetail'
    IssueListResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: object
          required:
          - items
          - nextCursor
          - hasMore
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Issue'
            nextCursor:
              type:
              - string
              - 'null'
            hasMore:
              type: boolean
...
    AnalyticsSummaryResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/AnalyticsSummary'
    RecurringResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/RecurringGroup'
    EffectivenessResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: '#/components/schemas/Effectiveness'
    EmployeeErrorStatsResponse:
      type: object
      required:
      - ok
      - data
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/EmployeeErrorStat'
...
  responses:
    BadRequest:
      description: Requête mal formée.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Non authentifié.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Forbidden:
      description: Interdit.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    NotFound:
      description: Introuvable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Conflict:
      description: Version périmée.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    ValidationError:
      description: Validation métier.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    PreconditionRequired:
      description: If-Match manquant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
    RateLimited:
      description: Limite de débit dépassée.
      headers:
        Retry-After:
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'